# GODRIVER-2526 Optimize description.selectByKind.