[GODRIVER-2526](https://jira.mongodb.org/browse/GODRIVER-2526)

The reporter has hundreds of mongos instances, so appending `Server` causes high memory usage. This PR pre-allocates slice to reduce allocations.

The benchmark show:
```
name                old time/op    new time/op    delta
Selector_Sharded-4    28.4µs ±37%     9.7µs ± 3%  -65.66%  (p=0.000 n=9+8)

name                old alloc/op   new alloc/op   delta
Selector_Sharded-4     111kB ± 0%      42kB ± 0%  -62.28%  (p=0.000 n=10+10)

name                old allocs/op  new allocs/op  delta
Selector_Sharded-4      9.00 ± 0%      3.00 ± 0%  -66.67%  (p=0.000 n=10+10)
```